feature_pipeline/src/requirements.txt
celery==5.3.1
pymongo==4.5.0
pika==1.3.2
python-dotenv==1.0.0
sentence-transformers==2.6.0
qdrant-client
torch==2.0.1
transformers==4.40.0
langchain_text_splitters
langchain-huggingface==0.2.0
langchain_qdrant
numpy==1.26.4